requirements.txt
joblib>=1.1.0
numba>=0.56.4
numpy>=1.23.5
scipy>=1.9.3
librosa>=0.9.1
llvmlite>=0.39.0
pydub>=0.25.1
soundfile>=0.12.1
ffmpeg-python>=0.2.0
tensorboardX
praat-parselmouth>=0.4.2
Pillow>=9.1.1
resampy>=0.4.2
scikit-learn
starlette>=0.25.0
tensorboard
tqdm>=4.63.1
audioread
torchcrepe>=0.0.20